# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel-Swagger-MockUpのAPIDoc",
        "description": "これはLaravel-Swagger-MockUpのAPIDocです。このドキュメントはdarkaonline/l5-swaggerを使用して生成されました。",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/create": {
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Create a category",
                "operationId": "46e16a82b4ad21ab764a45c638d6e55f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "required, max length 255",
                                        "type": "string",
                                        "example": "name1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/update/{id}": {
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Update a category",
                "operationId": "9962f68a913633dbbd530b863c6fe2c5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "required, max length 255",
                                        "type": "string",
                                        "example": "name1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "id": {
                        "description": "unique, incremented automatically",
                        "type": "int",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "description": "required, max length 255",
                        "type": "string",
                        "example": "name1"
                    },
                    "slug": {
                        "description": "required, max length 255",
                        "type": "string",
                        "example": "slug-1"
                    },
                    "created_at": {
                        "description": "not required, auto-generated",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2020-01-01 00:00:00"
                    },
                    "updated_at": {
                        "description": "not required, auto-generated",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2020-01-01 00:00:00"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}